Fix 16-bit masking bugs in 11 CPU instructions (CMP, SUB, SBB, SCASW, LOOP, CMPXCHG)

+++ crates/core/src/cpu_8086.rs
@@ -1848,10 +1848,10 @@ impl<M: Memory8086> Cpu8086<M> {
                     }
                     // SCASW
                     0xAF => {
-                        while self.cx != 0 {
+                        while (self.cx as u16) != 0 {
                             let val = self.read_u16(self.es, self.di as u16);
                             let result = (self.ax as u16).wrapping_sub(val);
-                            let borrow = self.ax < (val as u32);
+                            let borrow = (self.ax as u16) < val;
                             let overflow = (((self.ax as u16) ^ val)
                                 & ((self.ax as u16) ^ (result as u16))
                                 & 0x8000)
@@ -1980,10 +1980,10 @@ impl<M: Memory8086> Cpu8086<M> {
                     }
                     // SCASW
                     0xAF => {
-                        while self.cx != 0 {
+                        while (self.cx as u16) != 0 {
                             let val = self.read_u16(self.es, self.di as u16);
                             let result = (self.ax as u16).wrapping_sub(val);
-                            let borrow = self.ax < (val as u32);
+                            let borrow = (self.ax as u16) < val;
                             let overflow = (((self.ax as u16) ^ val)
                                 & ((self.ax as u16) ^ (result as u16))
                                 & 0x8000)
@@ -3065,7 +3065,7 @@ impl<M: Memory8086> Cpu8086<M> {
             0x2D => {
                 let val = self.fetch_u16();
                 let result = (self.ax as u16).wrapping_sub(val);
-                let borrow = self.ax < (val as u32);
+                let borrow = (self.ax as u16) < val;
                 let overflow =
                     (((self.ax as u16) ^ val) & ((self.ax as u16) ^ (result as u16)) & 0x8000) != 0;
                 let af = Self::calc_af_sub_16(self.ax as u16, val);
@@ -3100,7 +3100,7 @@ impl<M: Memory8086> Cpu8086<M> {
             0x3D => {
                 let val = self.fetch_u16();
                 let result = (self.ax as u16).wrapping_sub(val);
-                let borrow = self.ax < (val as u32);
+                let borrow = (self.ax as u16) < val;
                 let overflow =
                     (((self.ax as u16) ^ val) & ((self.ax as u16) ^ (result as u16)) & 0x8000) != 0;
                 let af = Self::calc_af_sub_16(self.ax as u16, val);
@@ -4083,7 +4083,7 @@ impl<M: Memory8086> Cpu8086<M> {
                         let result = self.ax.wrapping_sub(rm_val as u32);
                         self.set_flag(FLAG_SF, (result & 0x8000) != 0);
                         self.set_flag(FLAG_PF, (result & 0xFF).count_ones() % 2 == 0);
-                        let carry = self.ax < (rm_val as u32);
+                        let carry = (self.ax as u16) < rm_val;
                         let overflow = (((self.ax as u16) ^ rm_val)
                             & ((self.ax as u16) ^ (result as u16))
                             & 0x8000)
@@ -5081,7 +5081,7 @@ impl<M: Memory8086> Cpu8086<M> {
                 let val = self.fetch_u16();
                 let carry = if self.get_flag(FLAG_CF) { 1 } else { 0 };
                 let result = self.ax.wrapping_sub(val as u32).wrapping_sub(carry);
-                let borrow = self.ax < (val as u32 + carry as u32);
+                let borrow = (self.ax as u16) < (val + carry as u16);
                 let overflow =
                     (((self.ax as u16) ^ val) & ((self.ax as u16) ^ (result as u16)) & 0x8000) != 0;
                 // AF calculation: check if borrow from bit 4 to bit 3 in low byte including carry-in
@@ -6934,7 +6934,7 @@ impl<M: Memory8086> Cpu8086<M> {
             0xE0 => {
                 let offset = self.fetch_u8() as i8;
                 self.cx = self.cx.wrapping_sub(1);
-                if self.cx != 0 && !self.get_flag(FLAG_ZF) {
+                if (self.cx as u16) != 0 && !self.get_flag(FLAG_ZF) {
                     self.ip = (self.ip.wrapping_add((offset as i16) as u32)) & 0xFFFF;
                     self.cycles += 19;
                     19
@@ -6948,7 +6948,7 @@ impl<M: Memory8086> Cpu8086<M> {
             0xE1 => {
                 let offset = self.fetch_u8() as i8;
                 self.cx = self.cx.wrapping_sub(1);
-                if self.cx != 0 && self.get_flag(FLAG_ZF) {
+                if (self.cx as u16) != 0 && self.get_flag(FLAG_ZF) {
                     self.ip = (self.ip.wrapping_add((offset as i16) as u32)) & 0xFFFF;
                     self.cycles += 18;
                     18
@@ -6962,7 +6962,7 @@ impl<M: Memory8086> Cpu8086<M> {
             0xE2 => {
                 let offset = self.fetch_u8() as i8;
                 self.cx = self.cx.wrapping_sub(1);
-                if self.cx != 0 {
+                if (self.cx as u16) != 0 {
                     self.ip = (self.ip.wrapping_add((offset as i16) as u32)) & 0xFFFF;
                     self.cycles += 17;
                     17
@@ -6975,7 +6975,7 @@ impl<M: Memory8086> Cpu8086<M> {
             // JCXZ - Jump if CX is Zero (0xE3)
             0xE3 => {
                 let offset = self.fetch_u8() as i8;
-                if self.cx == 0 {
+                if (self.cx as u16) == 0 {
                     self.ip = (self.ip.wrapping_add((offset as i16) as u32)) & 0xFFFF;
                     self.cycles += 18;
                     18
@@ -7829,7 +7829,7 @@ impl<M: Memory8086> Cpu8086<M> {
                 // Compare AX with word at ES:DI
                 let val = self.read_u16(self.es, self.di as u16);
                 let result = (self.ax as u16).wrapping_sub(val);
-                let borrow = self.ax < (val as u32);
+                let borrow = (self.ax as u16) < val;
                 let overflow =
                     (((self.ax as u16) ^ val) & ((self.ax as u16) ^ (result as u16)) & 0x8000) != 0;
